Github ssh keys via example .ssh/config and .git/config files
The objective here is to push to github without requiring constantly
retyping the username and password. This is accomplished by setting up
an ssh key, referencin it in .ssh/config, and referencing this host
alias in .git/config, which can be accomplished by performing a ssh
clone on the repository using the alias defined in .ssh/config.

example_git_config is from a .git/config file from a repository folder.
example_ssh_config is from the .ssh/config file